+++ NEWS.md
@@ -1,5 +1,10 @@
 # cea 0.0.0.9000
 
+* Added `INMB()` & `INHB()` to calculate INMB/INHB from regression model
+  results.
+
+* Added `ICER()` to calculate ICER from regression model results.
+
 * Added `estimate()` to estimate CEA regression model.
 
 * Added a `NEWS.md` file to track changes to the package.